refactor: remove links with arguments in job form
The form for creating and updating job records is shared. In the event of creating a fresh form, there is no access to primary keys to identify a job to delete or return the the detailed view of. Although it would have made for good UX, the shared nature of the form makes this impossible.