feat: add safe filters enclosed in divs for services
The safe filter was required so that bullet points could be rendered from the template variable. The Summernote text will be rendered within p tags, whilst this was already contained in a p element on the template. By changing this to a div, the repeating p elements that HTML validators caught are no longer an issue, and the Summernote text can be rendered correctly.